:sparkles: feat : 독서 장소 리뷰 작성
feat : 독서 장소 리뷰 작성

+++ BE/src/main/java/com/seoro/seoro/domain/entity/Place/PlaceReview.java
@@ -27,7 +27,7 @@ public class PlaceReview implements Serializable {
     @ManyToOne(targetEntity = Member.class)
     @JoinColumn(name = "memberId")
     private Member member;
-    private Long score;
+    private Double score;
     private String reviewContent;
     @OneToMany(mappedBy = "placeReview")
     private List<PlaceReviewPhoto> photos = new ArrayList<>();
